+++ CHANGELOG.md
@@ -38,6 +38,7 @@ Release 5.8.0:
    - Extract `ValidatorMdoc`, `ValidatorSdJwt`, `ValidatorVcJws` from `Validator`
    - In `HolderAgent` add constructor parameters for `validatorVcJws`, `validatorSdJwt`, `validatorMdoc`
    - In `Validator` deprecate constructor parameter `resolveStatusListToken`, clients shall use `tokenStatusResolver` instead
+   - In `Verifier` remove parameter `challenge` from `verifyPresentationIsoMdoc()`
 
 Release 5.7.2:
  - Presentation Exchange: Fix validation of optional constraint fields